tres en raya: versión final

+++ tresenraya/application/models/gameModel.php
@@ -145,5 +145,6 @@ function checkWinner($filename)
 
 function resetGame($filename)
 {
-	unlink($filename);
+	if (file_exists($filename))
+		unlink($filename);
 }
